requirements.txt
tensorflow==2.10.1
pandas==2.3.0
numpy==1.24.0
fastapi==0.124.2
uvicorn==0.38.0
python-multipart==0.0.20
pillow==12.0.0
scikit-learn==1.7.0
